// packages/frontend/src/generated/types/ArrivalType.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const arrivalTypeEnum = {
  TRUCK: 'TRUCK',
  RAIL: 'RAIL',
  EXTERNAL: 'EXTERNAL',
  INITIAL_BALANCE: 'INITIAL_BALANCE',
} as const

export type ArrivalTypeEnumKey
  = (typeof arrivalTypeEnum)[keyof typeof arrivalTypeEnum]

export type ArrivalType = ArrivalTypeEnumKey