// packages/frontend/src/generated/types/CreateAcceptanceCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AcceptanceItemCompositeRequest } from './AcceptanceItemCompositeRequest.ts'
import type { CreateAcceptanceRequest } from './CreateAcceptanceRequest.ts'

export type CreateAcceptanceCompositeRequest = CreateAcceptanceRequest & {
  /**
   * @type array
   */
  items: AcceptanceItemCompositeRequest[]
}