// packages/frontend/src/generated/zod/acceptanceItemCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { AcceptanceItemCompositeRequest } from '../types/AcceptanceItemCompositeRequest.ts'
import * as z from 'zod'

export const acceptanceItemCompositeRequestSchema = z.object({
  acceptedAmount: z.string(),
  contractorId: z.string().uuid(),
  productId: z.string().uuid(),
  storageId: z.string().uuid(),
}) as unknown as ToZod<AcceptanceItemCompositeRequest>