// packages/frontend/src/generated/zod/createAcceptanceCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { CreateAcceptanceCompositeRequest } from '../types/CreateAcceptanceCompositeRequest.ts'
import * as z from 'zod'
import { acceptanceItemCompositeRequestSchema } from './acceptanceItemCompositeRequestSchema.ts'
import { createAcceptanceRequestSchema } from './createAcceptanceRequestSchema.ts'

export const createAcceptanceCompositeRequestSchema = z
  .lazy(() => createAcceptanceRequestSchema)
  .and(
    z.object({
      items: z.array(z.lazy(() => acceptanceItemCompositeRequestSchema)),
    }),
  ) as unknown as ToZod<CreateAcceptanceCompositeRequest>